+++ src/data/sqlite/dynamicSchema.js
@@ -4,6 +4,7 @@
 var execute = require('./execute'),
     promises = require('../../utilities/promises'),
     schemas = require('./schema'),
+    columns = require('./columns'),
     errorCodes = require('./errorCodes');
 
 module.exports = function (table) {
@@ -23,12 +24,15 @@ module.exports = function (table) {
                 if (attempt >= 3)
                     return promises.rejected(err);
 
-                // need to update column definition here
                 if(err.number === errorCodes.InvalidObjectName)
-                    return schema.createTable(table, item);
+                    return schema.createTable(table, item).then(updateColumns);
                 if(err.number === errorCodes.InvalidColumnName)
-                    return schema.updateSchema(table, item);
+                    return schema.updateSchema(table, item).then(updateColumns);
                 return promises.rejected(err);
+            }
+
+            function updateColumns() {
+                return columns.set(table, item);
             }
 
         },
