+++ src/data/sqlite/statements/createColumnsTable.js
